+++ CHANGELOG.md
@@ -33,6 +33,9 @@ Changelog
 * Add `elseif` as an alias of `elif` for parity with Twig. Thanks kswedberg.
   Merge of [#826](https://github.com/mozilla/nunjucks/pull/826).
 
+* Add nunjucks env to express app settings as `nunjucksEnv`.
+  Merge of [#829](https://github.com/mozilla/nunjucks/pull/829).
+
 * Add support for finding an object's "length" in length filter.
   Merge of [#813](https://github.com/mozilla/nunjucks/pull/813).